Start to isolate command handling in cmd.c
