+++ cmd.h
@@ -1,0 +1,6 @@
+#ifndef EMS_CMD_H
+#define EMS_CMD_H
+
+void cmd_title(int);
+
+#endif /* EMS_CMD_H */
